requirements.txt
streamlit
PyPDF2
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
